emotionless_options/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
yfinance>=0.2.0

# Data processing and analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0
ta-lib>=0.4.0

# Web scraping and API
requests>=2.26.0
beautifulsoup4>=4.9.0
aiohttp>=3.8.0

# Sentiment analysis
nltk>=3.6.0
textblob>=0.15.0
transformers>=4.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Configuration and logging
pyyaml>=6.0.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-asyncio>=0.16.0

# Development tools
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0 